solution_bot/requirements.txt
streamlit==1.24.0
langchain==0.0.228
openai==0.27.8
chromadb==0.3.26
python-docx==0.8.11
tiktoken==0.4.0
python-dotenv==1.0.0
PyPDF2==3.0.1
